+++ src/sssp.h
@@ -8,12 +8,12 @@ using namespace std;
 class PriceFunction {
     // Will be an integral price function
     public:
-    vector<int> prices;
+    vector<long long> prices;
     static PriceFunction sum(PriceFunction a, PriceFunction b);
 };
 
 SSSP_Result SPmain(Graph* g_in, int s_in);
-PriceFunction scaleDown(Graph* graph, int delta, int B, int depth);
+PriceFunction scaleDown(Graph* graph, long long delta, long long B, int depth);
 PriceFunction elimNeg(Graph* graph);
 PriceFunction FixDAGEdges(Graph* g, vector<set<int>>& v);
 pair<set<int>, set<pair<int, int>>> ballOut(Graph* graph, int v, int R, bool fromBallIn=false);
